+++ app/components/Postfeedback.tsx
@@ -1,34 +1,36 @@
-"use client";
-import { useRouter } from "next/navigation";
-import React, { useState } from "react";
-import { postFeedback } from "../hooks/useFeedback";
 import { Box, Button, TextField } from "@mui/material";
-import PostFeedbackUI from "./PostfeedbackUI";
+import React from "react";
 
-const PostFeedback = () => {
-  const router = useRouter();
-  const [feedback, setFeedback] = useState("");
+type PostFeedbackUIProps = {
+  feedback: string;
+  onFeedbackChange: (e: React.ChangeEvent<HTMLInputElement>) => void;
+  onSubmit: (e: React.FormEvent) => void;
+};
 
-  const handleSubmit = async (e: React.FormEvent) => {
-    e.preventDefault();
-
-    await postFeedback(feedback);
-    router.refresh();
-
-    setFeedback("");
-  };
-
-  const handleFeedbackChange = (e: React.ChangeEvent<HTMLInputElement>) => {
-    setFeedback(e.target.value);
-  };
-
+const PostFeedbackUI: React.FC<PostFeedbackUIProps> = ({
+  feedback,
+  onFeedbackChange,
+  onSubmit,
+}) => {
   return (
-    <PostFeedbackUI
-      feedback={feedback}
-      onFeedbackChange={handleFeedbackChange}
-      onSubmit={handleSubmit}
-    />
+    <Box component={"form"} onSubmit={onSubmit}>
+      <Box display={"flex"} flexDirection={"column"} mb={1}>
+        <TextField
+          sx={{ marginBottom: "20px" }}
+          value={feedback}
+          onChange={onFeedbackChange}
+          placeholder="感想を教えてね!"
+          multiline
+          minRows={3}
+        />
+      </Box>
+      <Box mb={3} justifyContent="center" display={"flex"}>
+        <Button variant="contained" type="submit" size="small">
+          投稿
+        </Button>
+      </Box>
+    </Box>
   );
 };
 
-export default PostFeedback;
+export default PostFeedbackUI;
